requirements.txt
Flask==3.1.1
python-telegram-bot==13.15
